day05

+++ src/main/java/life/suwei/community2/controller/PublishController.java
@@ -1,7 +1,6 @@
 package life.suwei.community2.controller;
 
 import life.suwei.community2.dto.QuestionDTO;
-import life.suwei.community2.mapper.QuestionMapper;
 import life.suwei.community2.model.Question;
 import life.suwei.community2.model.User;
 import life.suwei.community2.service.QuestionService;
@@ -27,7 +26,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,
+    public String edit(@PathVariable(name = "id")Long id,
                        Model model){
 
         QuestionDTO question = questionService.getById(id);
@@ -49,7 +48,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id",required = false)Integer id,
+            @RequestParam(value = "id",required = false)Long id,
             HttpServletRequest request,
             Model model
     ) {
